Enable multi statement tests. Also split long running tests into two test files
